Implemented getImage function
`getImage` now successfully returns a future `Uint8List` that can be passed directly to `Image.memory`.

Added this new functionality to the example app.

Changed API to take an ImageUri so that it matches the Spotify SDK. Also, majority of cases are likely to use `PlayerState.track.imageUri`, which is an ImageUri object.

Updated Dart to 2.7 so that I could use extension methods.