Implemented the PKCE authorization flow for the web platform. (#61)
Replaced the Implicit Grant authorization flow with the new Authorization Code With PKCE authorization flow. This allows the plugin to automatically refresh the authorization in the background every time the access token expires, instead of having to open the authorization prompt each time like before.
This requires a change to the web authentication redirect setup, therefore I updated the README with the new instructions.