Added ignore rule for unsafe html
Added ignore rule for unsafe html, as it should not be an issue for our use case.

As its meant really as an info prompt for the devs to be aware that they should not use user input to modify the HTML dom. 

There were a few open issues related to this warning, see: https://github.com/dart-lang/linter/issues/2348 & https://github.com/dart-lang/linter/issues/1604